pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "modelforge"
version = "0.1.0"
description = "Sistema de Fine-tuning de Modelos ML orientado por configuração YAML"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "ModelForge Team", email = "modelforge@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["machine-learning", "fine-tuning", "transformers", "huggingface", "nlp"]

dependencies = [
    "transformers>=4.36.0",
    "datasets>=2.16.0",
    "accelerate>=0.25.0",
    "huggingface-hub>=0.20.0",
    "torch>=2.1.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "flask>=3.0.0",
    "gunicorn>=21.2.0",
    "docker>=7.0.0",
    "structlog>=24.1.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.26.0",
    "evaluate>=0.4.0",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
]

[project.scripts]
modelforge = "modelforge.cli:main"

[project.urls]
Homepage = "https://github.com/modelforge/modelforge"
Documentation = "https://github.com/modelforge/modelforge#readme"
Repository = "https://github.com/modelforge/modelforge"
Issues = "https://github.com/modelforge/modelforge/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["modelforge*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
modelforge = ["templates/*.yaml", "templates/*.yml", "templates/Dockerfile"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["modelforge"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501", # line too long (handled by black)
]

